+++ src/gui/src/widgets/Sidebar.cpp
@@ -44,39 +44,49 @@ void Sidebar::initLayout()
 void Sidebar::addBottomControls()
 {
     m_layout->addItem(new QSpacerItem(20, 40, QSizePolicy::Minimum, QSizePolicy::Expanding));
+    addButtonSettings();
     addButtonLogs();
     addButtonHelp();
+}
+
+void Sidebar::addButtonSettings()
+{
+    QIcon icon(QString::fromUtf8(":/res/icons/sidebar/settings.svg"));
+    auto button = addButton(icon);
+    connect(button, &QPushButton::clicked, this, [this]() {
+        emit OpenSettings();
+    });
 }
 
 void Sidebar::addButtonLogs()
 {
     QIcon icon(QString::fromUtf8(":/res/icons/sidebar/logs.svg"));
-    auto button = new QPushButton(this);
-    button->setFlat(true);
-    button->setIcon(icon);
-    button->setIconSize(QSize(30, 30));
-    button->setCursor(Qt::PointingHandCursor);
+    auto button = addButton(icon);
     connect(button, &QPushButton::clicked, this, [this]() {
         emit OpenLogs();
     });
-    m_layout->addWidget(button);
 }
 
 void Sidebar::addButtonHelp()
 {
     QIcon icon(QString::fromUtf8(":/res/icons/sidebar/help.svg"));
+    auto button = addButton(icon);
+    connect(button, &QPushButton::clicked, []() {
+        const QUrl documentation("https://github.com/SergeyGadzhilov/SKVM/wiki");
+        QDesktopServices::openUrl(documentation);
+    });
+}
 
+QPushButton* Sidebar::addButton(QIcon icon)
+{
     auto button = new QPushButton(this);
     button->setFlat(true);
     button->setIcon(icon);
     button->setIconSize(QSize(30, 30));
     button->setCursor(Qt::PointingHandCursor);
-    connect(button, &QPushButton::clicked, []() {
-        const QUrl documentation("https://github.com/SergeyGadzhilov/SKVM/wiki");
-        QDesktopServices::openUrl(documentation);
-    });
-
     m_layout->addWidget(button);
+
+    return button;
 }
 
 } //namespace skvm_widgets